# perf(rust, python): parallelize numeric ChunkedArray materialization `~2x`.