This code is hit a lot, so this is a huge win.

Tested on null densities: `0.0, 0.2, 0.5, 0.8` percent. 

50% null density is worst as expected as that kills predictability

# Master branch:

```
[]  ritchie46 /home/ritchie46/code/memcheck$ ./target/release/base                            (base) 
null density: 0 took 441ms
100000000
null density: 0.2 took 607ms
100000000
null density: 0.5 took 905ms
100000000
null density: 0.8 took 604ms
100000000
[src/main.rs:3] group_tuples::run() = ()
[]  ritchie46 /home/ritchie46/code/memcheck$ ./target/release/base                            (base) 
null density: 0 took 417ms
100000000
null density: 0.2 took 594ms
100000000
null density: 0.5 took 854ms
100000000
null density: 0.8 took 574ms
100000000
[src/main.rs:3] group_tuples::run() = ()
[]  ritchie46 /home/ritchie46/code/memcheck$ ./target/release/base                            (base) 
null density: 0 took 431ms
100000000
null density: 0.2 took 569ms
100000000
null density: 0.5 took 887ms
100000000
null density: 0.8 took 579ms
100000000
```

# this PR

```
[src/main.rs:3] group_tuples::run() = ()
[]  ritchie46 /home/ritchie46/code/memcheck$ ./target/release/memcheck                        (base) 
null density: 0 took 193ms
100000000
null density: 0.2 took 367ms
100000000
null density: 0.5 took 488ms
100000000
null density: 0.8 took 325ms
100000000
[src/main.rs:3] group_tuples::run() = ()
[]  ritchie46 /home/ritchie46/code/memcheck$ ./target/release/memcheck                        (base) 
null density: 0 took 190ms
100000000
null density: 0.2 took 308ms
100000000
null density: 0.5 took 468ms
100000000
null density: 0.8 took 366ms
100000000
[src/main.rs:3] group_tuples::run() = ()
[]  ritchie46 /home/ritchie46/code/memcheck$ ./target/release/memcheck                        (base) 
null density: 0 took 195ms
100000000
null density: 0.2 took 339ms
100000000
null density: 0.5 took 490ms
100000000
null density: 0.8 took 357ms
100000000
[src/main.rs:3] group_tuples::run() = ()
```